requirements.txt
Flask
flask-cors
google-genai
python-dotenv
gunicorn
flask-sqlalchemy
psycopg2-binary
